Add tests for WatchProjectViewModel
