+++ KsApi/models/GraphMutationWatchProjectResponseEnvelope.swift
@@ -1,14 +1,14 @@
 import Foundation
 
 public struct GraphMutationWatchProjectResponseEnvelope: Decodable {
-  public let watchProject: WatchProject
+  public private(set) var watchProject: WatchProject
 
   public struct WatchProject: Decodable {
-    public let project: Project
+    public private(set) var project: Project
 
     public struct Project: Decodable {
-      public let id: String
-      public let isWatched: Bool
+      public private(set) var id: String
+      public private(set) var isWatched: Bool
     }
   }
 }
