+++ KsApi/models/templates/GraphMutationUnwatchProjectResponseEnvelopeTemplates.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension GraphMutationUnwatchProjectResponseEnvelope {
+  internal static let template = GraphMutationUnwatchProjectResponseEnvelope(
+    unwatchProject: .init(
+      project: .init(id: "UHJvamVjdC0xMzEzNzE3MDgy", isWatched: false)
+    )
+  )
+}
