+++ KsApi/models/templates/GraphMutationWatchProjectResponseEnvelopeTemplates.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension GraphMutationWatchProjectResponseEnvelope {
+  internal static let template = GraphMutationWatchProjectResponseEnvelope(
+    watchProject: .init(
+      project: .init(id: "UHJvamVjdC0xMzEzNzE3MDgy", isWatched: true)
+    )
+  )
+}
