+++ Library/ViewModels/ProjectNavBarViewModelTests.swift
@@ -17,13 +17,7 @@ final class ProjectNavBarViewModelTests: TestCase {
   fileprivate let categoryHidden = TestObserver<Bool, NoError>()
   fileprivate let categoryAnimate = TestObserver<Bool, NoError>()
   fileprivate let dismissViewController = TestObserver<(), NoError>()
-  fileprivate let generateSelectionFeedback = TestObserver<(), NoError>()
-  fileprivate let generateSuccessFeedback = TestObserver<(), NoError>()
-  fileprivate let goToLoginTout = TestObserver<(), NoError>()
   fileprivate let projectName = TestObserver<String, NoError>()
-  fileprivate let saveButtonSelected = TestObserver<Bool, NoError>()
-  fileprivate let showProjectSavedPrompt = TestObserver<Void, NoError>()
-  fileprivate let saveButtonAccessibilityValue = TestObserver<String, NoError>()
   fileprivate let titleAnimate = TestObserver<Bool, NoError>()
   fileprivate let titleHidden = TestObserver<Bool, NoError>()
 
@@ -37,13 +31,7 @@ final class ProjectNavBarViewModelTests: TestCase {
     self.vm.outputs.categoryHiddenAndAnimate.map(first).observe(self.categoryHidden.observer)
     self.vm.outputs.categoryHiddenAndAnimate.map(second).observe(self.categoryAnimate.observer)
     self.vm.outputs.dismissViewController.observe(self.dismissViewController.observer)
-    self.vm.outputs.generateSelectionFeedback.observe(self.generateSelectionFeedback.observer)
-    self.vm.outputs.generateSuccessFeedback.observe(self.generateSuccessFeedback.observer)
-    self.vm.outputs.goToLoginTout.observe(self.goToLoginTout.observer)
     self.vm.outputs.projectName.observe(self.projectName.observer)
-    self.vm.outputs.showProjectSavedPrompt.observe(self.showProjectSavedPrompt.observer)
-    self.vm.outputs.saveButtonSelected.observe(self.saveButtonSelected.observer)
-    self.vm.outputs.saveButtonAccessibilityValue.observe(self.saveButtonAccessibilityValue.observer)
     self.vm.outputs.titleHiddenAndAnimate.map(second).observe(self.titleAnimate.observer)
     self.vm.outputs.titleHiddenAndAnimate.map(first).observe(self.titleHidden.observer)
   }
@@ -173,252 +161,6 @@ final class ProjectNavBarViewModelTests: TestCase {
     self.vm.inputs.closeButtonTapped()
 
     self.dismissViewController.assertValueCount(1)
-  }
-
-  func testGenerateSelectionFeedback() {
-    let project = .template |> Project.lens.personalization.isStarred .~ false
-
-    self.vm.inputs.configureWith(project: project, refTag: nil)
-    self.vm.inputs.viewDidLoad()
-    self.vm.inputs.saveButtonTapped(selected: true)
-    self.scheduler.advance()
-    self.generateSelectionFeedback.assertValueCount(1)
-    self.generateSuccessFeedback.assertValueCount(0)
-  }
-
-  func testGenerateSuccessFeedback() {
-    let project = .template |> Project.lens.personalization.isStarred .~ false
-
-    self.vm.inputs.configureWith(project: project, refTag: nil)
-    self.vm.inputs.viewDidLoad()
-    self.vm.inputs.saveButtonTapped(selected: false)
-    self.scheduler.advance()
-    self.generateSelectionFeedback.assertValueCount(0)
-    self.generateSuccessFeedback.assertValueCount(1)
-  }
-
-  // Tests the flow of a logged out user trying to save a project, and then going through the login flow.
-  func testLoggedOutUser_SavesProject() {
-    let project = .template |> Project.lens.personalization.isStarred .~ false
-    let toggleSaveResponse = .template
-      |> StarEnvelope.lens.project .~ (project |> Project.lens.personalization.isStarred .~ true)
-
-    withEnvironment(apiService: MockService(toggleStarResponse: toggleSaveResponse)) {
-      self.saveButtonSelected.assertDidNotEmitValue("No values emitted at first.")
-      self.vm.inputs.configureWith(project: project, refTag: nil)
-      self.vm.inputs.viewDidLoad()
-
-      self.saveButtonSelected.assertValues([false], "Save button is not selected at first")
-
-      self.vm.inputs.saveButtonTapped(selected: true)
-
-      self.saveButtonSelected.assertValues([false],
-                                            "Nothing is emitted when save button tapped while logged out.")
-      self.saveButtonAccessibilityValue.assertValues(["Unsaved"])
-
-      self.goToLoginTout.assertValueCount(1, "Prompt to login when saving while logged out.")
-
-      AppEnvironment.login(.init(accessToken: "deadbeef", user: .template))
-      self.vm.inputs.userSessionStarted()
-
-      self.saveButtonSelected.assertValues([false, true],
-                                            "Once logged in, the save button selects immediately.")
-      self.saveButtonAccessibilityValue.assertValues(["Unsaved", "Saved"])
-
-      self.scheduler.advance(by: .seconds(1))
-
-      self.saveButtonSelected.assertValues([false, true],
-                                           "Save button stays selected after API request.")
-      self.showProjectSavedPrompt.assertValueCount(0, "The save project prompt does not show.")
-      XCTAssertEqual(["Project Star", "Starred Project", "Saved Project"],
-                     trackingClient.events, "A star koala event is tracked.")
-    }
-  }
-
-  // Tests a logged in user saving a project.
-  func testLoggedInUser_SavesAndUnsavesProject() {
-    AppEnvironment.login(.init(accessToken: "deadbeef", user: .template))
-
-    let project = Project.template
-    let toggleSaveResponse = .template
-      |> StarEnvelope.lens.project .~ (project |> Project.lens.personalization.isStarred .~ true)
-
-    withEnvironment(apiService: MockService(toggleStarResponse: toggleSaveResponse)) {
-      self.vm.inputs.configureWith(project: project, refTag: nil)
-      self.vm.inputs.viewDidLoad()
-
-      self.saveButtonSelected.assertValues([false], "Save button is not selected at first")
-      self.saveButtonAccessibilityValue.assertValues(["Unsaved"])
-
-      self.vm.inputs.saveButtonTapped(selected: true)
-
-      self.saveButtonSelected.assertValues([false, true], "Save button selects immediately.")
-      self.saveButtonAccessibilityValue.assertValues(["Unsaved", "Saved"])
-
-      self.scheduler.advance(by: .seconds(1))
-
-      self.saveButtonSelected.assertValues([false, true],
-                                           "Save button remains selected.")
-
-      self.showProjectSavedPrompt.assertValueCount(1, "The save project prompt shows.")
-      XCTAssertEqual(["Project Star", "Starred Project", "Saved Project"],
-                     trackingClient.events, "A star koala event is tracked.")
-
-      let untoggleSaveResponse = .template
-        |> StarEnvelope.lens.project .~ (project |> Project.lens.personalization.isStarred .~ false)
-
-      withEnvironment(apiService: MockService(toggleStarResponse: untoggleSaveResponse)) {
-        self.vm.inputs.saveButtonTapped(selected: true)
-
-        self.saveButtonSelected.assertValues([false, true, false],
-                                             "Save button deselects immediately.")
-        self.saveButtonAccessibilityValue.assertValues(["Unsaved", "Saved", "Unsaved"])
-
-        self.scheduler.advance(by: .seconds(1))
-
-        self.saveButtonSelected.assertValues([false, true, false],
-                                             "The save button remains unselected.")
-        self.saveButtonAccessibilityValue.assertValues(["Unsaved", "Saved", "Unsaved"])
-
-        self.showProjectSavedPrompt.assertValueCount(1, "The save project prompt only showed for starring.")
-        XCTAssertEqual(["Project Star", "Starred Project", "Saved Project", "Project Unstar",
-                        "Unstarred Project", "Unsaved Project"],
-                       self.trackingClient.events, "An unstar koala event is tracked.")
-      }
-    }
-  }
-
-  func testLoggedInUser_SaveProjectDebouncing() {
-    AppEnvironment.login(.init(accessToken: "deadbeef", user: .template))
-
-    let project = Project.template
-    let toggleSaveResponse = .template
-      |> StarEnvelope.lens.project .~ (project |> Project.lens.personalization.isStarred .~ true)
-
-    withEnvironment(apiService: MockService(toggleStarResponse: toggleSaveResponse)) {
-      self.vm.inputs.configureWith(project: project, refTag: nil)
-      self.vm.inputs.viewDidLoad()
-
-      self.saveButtonSelected.assertValues([false], "Save button is not selected at first")
-      self.saveButtonAccessibilityValue.assertValues(["Unsaved"])
-
-      self.vm.inputs.saveButtonTapped(selected: true)
-
-      self.saveButtonSelected.assertValues([false, true], "Save button selects immediately.")
-      self.saveButtonAccessibilityValue.assertValues(["Unsaved", "Saved"])
-
-      self.scheduler.advance(by: .milliseconds(500))
-
-      self.vm.inputs.saveButtonTapped(selected: true)
-
-      self.scheduler.advance(by: .milliseconds(500))
-
-      self.vm.inputs.saveButtonTapped(selected: true)
-
-      self.scheduler.advance(by: .milliseconds(500))
-
-      self.vm.inputs.saveButtonTapped(selected: true)
-
-      self.saveButtonSelected.assertValues(
-        [false, true, false, true, false],
-        "State flips back and forth as button is tapped."
-      )
-
-      self.scheduler.advance(by: .seconds(1))
-
-      self.saveButtonSelected.assertValues(
-        [false, true, false, true, false, true],
-        "Network call is made after 1 second of no taps, value from response is always used"
-      )
-
-      self.showProjectSavedPrompt.assertValueCount(1, "The save project prompt shows.")
-      XCTAssertEqual(["Project Star", "Starred Project", "Saved Project"],
-                     trackingClient.events, "A star koala event is tracked.")
-    }
-  }
-
-  // Tests a logged in user saving a project that ends soon.
-  func testLoggedInUser_SaveEndingSoonProject() {
-    AppEnvironment.login(.init(accessToken: "deadbeef", user: .template))
-
-    let project = .template
-      |> Project.lens.personalization.isStarred .~ false
-      |> Project.lens.dates.deadline .~ (MockDate().date.timeIntervalSince1970 + 60.0 * 60.0 * 24.0)
-
-    let toggleSaveResponse = .template
-      |> StarEnvelope.lens.project .~ (project |> Project.lens.personalization.isStarred .~ true)
-
-    withEnvironment(apiService: MockService(toggleStarResponse: toggleSaveResponse)) {
-      self.vm.inputs.configureWith(project: project, refTag: nil)
-      self.vm.inputs.viewDidLoad()
-      self.vm.inputs.saveButtonTapped(selected: true)
-      self.scheduler.advance(by: .seconds(1))
-
-      self.showProjectSavedPrompt.assertValueCount(
-        0, "The save project prompt doesn't show cause it's less than 48hrs."
-      )
-
-      XCTAssertEqual(["Project Star", "Starred Project", "Saved Project"], self.trackingClient.events,
-                     "A star koala event is tracked.")
-    }
-  }
-
-  // Tests a user unsaving a project.
-  func testLoggedInUser_UnsavesProject() {
-    AppEnvironment.login(.init(accessToken: "deadbeef", user: .template))
-
-    let project = .template |> Project.lens.personalization.isStarred .~ true
-    let toggleSaveResponse = .template
-      |> StarEnvelope.lens.project .~ (project |> Project.lens.personalization.isStarred .~ false)
-
-    withEnvironment(apiService: MockService(toggleStarResponse: toggleSaveResponse)) {
-      self.vm.inputs.configureWith(project: project, refTag: nil)
-      self.vm.inputs.viewDidLoad()
-      self.vm.inputs.saveButtonTapped(selected: true)
-      self.scheduler.advance(by: .seconds(1))
-
-      self.showProjectSavedPrompt.assertValueCount(0, "The save project prompt does not show.")
-
-      XCTAssertEqual(["Project Unstar", "Unstarred Project", "Unsaved Project"], self.trackingClient.events,
-                     "An unstar koala event is tracked.")
-    }
-  }
-
-  func testLoggedInSaveFailure() {
-    /// CHECK THIS
-    let error = ErrorEnvelope(
-      errorMessages: ["Something went wrong."],
-      ksrCode: .UnknownCode,
-      httpCode: 404,
-      exception: nil
-    )
-
-    let project = .template |> Project.lens.personalization.isStarred .~ false
-
-    withEnvironment(apiService: MockService(toggleStarError: error), currentUser: .template) {
-      self.vm.inputs.configureWith(project: project, refTag: nil)
-
-      self.vm.inputs.viewDidLoad()
-      self.vm.inputs.saveButtonTapped(selected: true)
-
-      self.saveButtonSelected.assertValues([false, true])
-
-      self.scheduler.advance(by: .seconds(1))
-
-      self.saveButtonSelected.assertValues([false, true, false])
-
-      self.showProjectSavedPrompt.assertValueCount(1, "The save project prompt shows.")
-      XCTAssertEqual([], trackingClient.events, "The star event does not track.")
-
-      self.vm.inputs.saveButtonTapped(selected: true)
-    }
-
-    // fix:
-    //    self.starButtonSelected.assertValues([false, true, false, true])
-    //
-    //    self.scheduler.advance()
-    //
-    //    self.starButtonSelected.assertValues([false, true, false, true, false])
   }
 
   func testTitleHiddenAndAnimate() {
